update detail and api add icon application
