+++ app/src/main/java/com/example/foodictive/MainActivity.kt
@@ -3,33 +3,20 @@ package com.example.foodictive
 import android.Manifest
 import android.content.Intent
 import android.content.pm.PackageManager
-import android.graphics.BitmapFactory
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.view.GestureDetector
-import android.view.MotionEvent
 import android.widget.Toast
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
 import com.example.foodictive.databinding.ActivityMainBinding
 import com.example.foodictive.response.FoodsItem
-import com.example.foodictive.view.CameraActivity
 import com.example.foodictive.view.detail.DetailMakanan
 import com.google.firebase.auth.FirebaseAuth
-import java.io.File
-import kotlin.math.abs
 
-class MainActivity : AppCompatActivity(), GestureDetector.OnGestureListener {
+class MainActivity : AppCompatActivity(){
 
     private lateinit var binding: ActivityMainBinding
     private lateinit var firebaseAuth: FirebaseAuth
-    private var getFile: File? = null
-    private lateinit var gestureDetector: GestureDetector
-    private var x1 = 0.0f
-    private var x2 = 0.0f
-    private var y1 = 0.0f
-    private var y2 = 0.0f
     private lateinit var food: FoodsItem
 
     override fun onRequestPermissionsResult(
@@ -63,57 +50,13 @@ class MainActivity : AppCompatActivity(), GestureDetector.OnGestureListener {
         }
         binding.halamanDetail.setOnClickListener { intentDetail(food) }
 
-        gestureDetector = GestureDetector(this,this)
     }
 
     private fun intentDetail(food: FoodsItem){
             val intent = Intent(this,DetailMakanan::class.java)
             intent.putExtra("food",food)
             startActivity(intent)
-
     }
-
-
-    override fun onTouchEvent(event: MotionEvent?): Boolean {
-        gestureDetector.onTouchEvent(event)
-
-        when(event?.action){
-            0->{
-                x1 = event.x
-                y1 = event.y
-            }
-            1->{
-                x2 = event.x
-                y2 = event.y
-
-                val valueX = x2-x1
-
-                if (abs(valueX) > MIN_DISTANCE){
-                    if (x2 > x1){
-                        val intent = Intent(this,CameraActivity::class.java)
-                        intentCamera.launch(intent)
-                        overridePendingTransition(R.anim.slide_in_left,R.anim.slide_out_right)
-                    }
-                }
-            }
-        }
-        return super.onTouchEvent(event)
-    }
-
-    private val intentCamera = registerForActivityResult(
-        ActivityResultContracts.StartActivityForResult()
-    ){
-        if (it.resultCode == CAMERA_RESULT){
-            val myFile = it.data?.getSerializableExtra("picture") as File
-
-            getFile = myFile
-            val result = BitmapFactory.decodeFile(myFile.path)
-
-            binding.previewImageView.setImageBitmap(result)
-        }
-    }
-
-
 
     private fun chckUser() {
         val firebaseUser = firebaseAuth.currentUser
@@ -127,33 +70,8 @@ class MainActivity : AppCompatActivity(), GestureDetector.OnGestureListener {
         }
     }
 
-
-
-    override fun onDown(p0: MotionEvent?): Boolean {
-        return false
-    }
-
-    override fun onShowPress(p0: MotionEvent?) {
-    }
-
-    override fun onSingleTapUp(p0: MotionEvent?): Boolean {
-        return false
-    }
-
-    override fun onScroll(p0: MotionEvent?, p1: MotionEvent?, p2: Float, p3: Float): Boolean {
-        return false
-    }
-
-    override fun onLongPress(p0: MotionEvent?) {
-    }
-
-    override fun onFling(p0: MotionEvent?, p1: MotionEvent?, p2: Float, p3: Float): Boolean {
-        return false
-    }
-
     companion object{
         const val CAMERA_RESULT = 200
-        const val MIN_DISTANCE = 150
         private val REQUIRED_PERMISSION = arrayOf(Manifest.permission.CAMERA)
         private const val REQUEST_CODE_PERMISSION = 10
     }
