+++ app/src/main/java/com/example/foodictive/api/apiConfig.kt
@@ -12,7 +12,7 @@ class apiConfig {
             val loggingInterceptor = if (BuildConfig.DEBUG) HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY)
             else HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.NONE)
             val client = OkHttpClient.Builder().addInterceptor(loggingInterceptor).build()
-            val retrofit = Retrofit.Builder().baseUrl("http://34.87.172.25:5000/")
+            val retrofit = Retrofit.Builder().baseUrl("http://34.87.132.191:5000/")
                 .addConverterFactory(GsonConverterFactory.create())
                 .client(client)
                 .build()
