+++ app/src/main/java/com/example/foodictive/view/detail/DetailMakanan.kt
@@ -2,17 +2,21 @@ package com.example.foodictive.view.detail
 
 import android.content.Intent
 import android.graphics.Bitmap
+import android.graphics.BitmapFactory
 import android.net.Uri
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.provider.MediaStore
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.lifecycle.ViewModelProvider
+import com.example.foodictive.MainActivity.Companion.CAMERA_RESULT
 import com.example.foodictive.MapsActivity
+import com.example.foodictive.R
 import com.example.foodictive.databinding.ActivityDetailMakananBinding
 import com.example.foodictive.ml.ModelFp16
 import com.example.foodictive.response.FoodsItem
 import com.example.foodictive.uriToFile
+import com.example.foodictive.view.CameraActivity
 import org.tensorflow.lite.support.image.TensorImage
 import java.io.File
 
@@ -22,6 +26,7 @@ class DetailMakanan : AppCompatActivity() {
     private var getFile: File? = null
     private lateinit var bitmap:Bitmap
 
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityDetailMakananBinding.inflate(layoutInflater)
@@ -29,6 +34,9 @@ class DetailMakanan : AppCompatActivity() {
         binding.identifikasi.setOnClickListener { setupViewModel() }
         binding.addGalery.setOnClickListener { startGalery() }
         binding.showMap.setOnClickListener{ startMaps() }
+        binding.tambahFoto.setOnClickListener { val intent = Intent(this, CameraActivity::class.java)
+            intentCamera.launch(intent)
+            overridePendingTransition(R.anim.slide_in_left, R.anim.slide_out_right) }
 
     }
 
@@ -53,6 +61,20 @@ class DetailMakanan : AppCompatActivity() {
             getFile = myFile
             binding.imageDetail.setImageURI(selectedImg)
             bitmap = MediaStore.Images.Media.getBitmap(this.contentResolver,selectedImg)
+        }
+    }
+
+    private val intentCamera = registerForActivityResult(
+        ActivityResultContracts.StartActivityForResult()
+    ){
+        if (it.resultCode == CAMERA_RESULT){
+            val myFile = it.data?.getSerializableExtra("picture") as File
+
+            getFile = myFile
+            val result = BitmapFactory.decodeFile(myFile.path)
+
+            binding.imageDetail.setImageBitmap(result)
+            bitmap = result
         }
     }
